fix: harden PATCH /users/id against broken access control (#95)
* fix: align user service tests with updated PATCH implementation

* feat: add seed script for PATCH users

* feat: add test script for PATCH /users/{id}

* fix: use BigDecimal for purchase_price to match DB NUMERIC

* fix: remove email from PATCH /users/{id}

* fix: remove email from UserUpdateRequest schema in OpenAPI spec

* fix: re-fetch user after patch to avoid Hibernate L1 cache staleness

* fix: use real tokens in patch user test script

* style: apply formatting

+++ src/main/java/hei/school/library/dto/ArrivalBookCopyRequest.java
@@ -1,5 +1,6 @@
 package hei.school.library.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -10,6 +11,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class ArrivalBookCopyRequest {
   private UUID bookCopyId;
-  private Double purchasePrice;
+  private BigDecimal purchasePrice;
   private Integer quantity;
 }
